# config/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "SECOND HAND APPLICATION",
    "version": "1.0",
    "contact": {
      "name": "Fadli-Devi-Aulia",
      "url": "https://github.com/wly0601/fp-fsw-backend.git",
      "email": "secondhand@gmail.com"
    }
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/transaction": {
      "post": {
        "summary": "Create Transaction",
        "description": "Add new transaction on the database",
        "operationId": "createTransaction",
        "security": [{
          "BearerTokenAuth": []
        }],
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/coreTransaction"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [{
          "name": "body",
          "in": "body",
          "required": true,
          "schema": {
            "$ref": "#/definitions/coreCreateTransactionRequest"
          }
        }],
        "tags": [
          "Transaction"
        ]
      }
    },
...
    "/api/transaction/{id}": {
      "get": {
        "summary": "Get Transaction By Id",
        "description": "Get Transaction By Id ",
        "security": [{
          "BearerTokenAuth": []
        }],
        "responses": {
          "200": {
            "description": "Berhasil ditampilkan",
            "content": {
              "application/json": {}
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [{
          "name": "id",
          "in": "path",
          "required": true,
          "type": "integer",
          "format": "int64"
        }],
        "tags": [
          "Transaction"
        ]
      },
      "put": {
        "summary": "Update Transaction",
        "description": "Update Transaction",
        "operationId": "updateTransaction",
        "security": [{
          "BearerTokenAuth": []
        }],
        "responses": {
          "200": {
            "description": "A successful response."
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [{
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "accBySeller": {
                  "type": "boolean"
                }
              }
            }
          }
        ],
        "tags": [
          "Transaction"
        ]
      }
    },
    "/api/transaction/{id}/confirmation": {
      "put": {
        "summary": "Update Confirmation Seller",
        "description": "Update Confirmation Seller to Cancel or SoldOut",
        "operationId": "confirmationSeller",
        "security": [{
          "BearerTokenAuth": []
        }],
        "responses": {
          "200": {
            "description": "A successful response."
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [{
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "isCanceled": {
                  "type": "boolean"
                }
              }
            }
          }
        ],
        "tags": [
          "Transaction"
        ]
      }
    },
...
    "/api/product/{id}": {
      "get": {
        "summary": "Get Product",
        "description": "Retrieve product detail",
        "operationId": "getProduct",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/coreProduct"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [{
          "name": "id",
          "in": "path",
          "required": true,
          "type": "integer",
          "format": "int64"
        }],
        "tags": [
          "Products"
        ]
      },
      "delete": {
        "summary": "Delete Product",
        "description": "Remove existing product by id",
        "operationId": "deleteProduct",
        "security": [{
          "BearerTokenAuth": []
        }],
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "properties": {}
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [{
          "name": "id",
          "in": "path",
          "required": true,
          "type": "integer",
          "format": "int64"
        }],
        "tags": [
          "Products"
        ]
      },
      "put": {
        "summary": "Update Product",
        "description": "Change product information",
        "operationId": "updateProduct",
        "security": [{
          "BearerTokenAuth": []
        }],
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/coreProduct"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [{
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "price": {
                  "type": "integer"
                },
                "categoryId": {
                  "type": "integer"
                },
                "description": {
                  "type": "string"
                },
                "images": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        ],
        "tags": [
          "Products"
        ]
      }
    }
  },
  "definitions": {
...
    "coreProduct": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "name": {
          "type": "string"
        },
        "price": {
          "type": "integer"
        },
        "categoryId": {
          "type": "integer"
        },
        "description": {
          "type": "string"
        },
        "images": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "coreTransaction": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "productId": {
          "type": "integer"
        },
        "buyerId": {
          "type": "integer"
        },
        "bargainPrice": {
          "type": "integer"
        },
        "dateOfBargain": {
          "type": "string",
          "format": "date-time"
        },
        "accBySeller": {
          "type": "boolean"
        },
        "dateOfAccOrNot": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "coreCreateTransactionRequest": {
      "type": "object",
      "properties": {
        "productId": {
          "type": "integer"
        },
        "bargainPrice": {
          "type": "integer"
        }
      }
    },
...
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  },
  "securityDefinitions": {
    "BearerTokenAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization"
    }
  }
}